Moved real-time function into a single module
